refactoring conversatoins
